+++ src/app/authorize/authorize.module.ts
@@ -12,10 +12,12 @@ import { A11yLinkModule } from '../cdk/a11y-link/a11y-link.module'
 import { InfoDropDownModule } from '../cdk/info-drop-down/info-drop-down.module'
 import { TrustedIndividualsDropdownModule } from '../cdk/trusted-individuals-dropdown/trusted-individuals-dropdown.module'
 import { AuthorizeRoutingModule } from './authorize-routing.module'
-import { AuthorizeComponent } from './pages/authorize/authorize.component'
+import { OauthErrorComponent } from './components/oauth-error/oauth-error.component'
+import { AuthorizeComponent } from './pages/authorize/authorize.component';
+import { FormAuthorizeComponent } from './components/form-authorize/form-authorize.component'
 
 @NgModule({
-  declarations: [AuthorizeComponent],
+  declarations: [AuthorizeComponent, OauthErrorComponent, FormAuthorizeComponent],
   imports: [
     CommonModule,
     AuthorizeRoutingModule,
