+++ src/app/authorize/components/oauth-error/oauth-error.component.html
@@ -7,6 +7,6 @@
     i18n="@@authorize.authorizeAccess"
   />
   <p class="mat-body-1">
-    {{ errorDescription }}
+    {{ error }}
   </p>
 </div>
