+++ src/app/authorize/components/oauth-error/oauth-error.component.scss
