+++ src/app/authorize/pages/authorize/authorize.component.html
@@ -1,166 +1,6 @@
 <main id="main">
   <div class="container">
-    <div class="row space-around">
-      <div class="col l5 m6 s4">
-        <mat-card
-          class="orcid-wizard-2"
-          *ngIf="oauthRequest && userName && orcidUrl"
-        >
-          <mat-card-header>
-            <img
-              class="logo mat-card-logo"
-              src="./assets/vectors/orcid.logo.svg"
-              alt="orcid logo"
-              role="link"
-              i18n="@@authorize.authorizeAccess"
-            />
-            <mat-card-title i18n="@@authorize.authorize"
-              >Authorize access</mat-card-title
-            >
-          </mat-card-header>
-          <mat-progress-bar
-            *ngIf="loadingUserInfo || loadingTrustedIndividuals"
-            mode="indeterminate"
-          ></mat-progress-bar>
-          <mat-card-content>
-            <mat-divider class="orcid-divider"></mat-divider>
-            <p
-              class="mat-caption m-b-16 m-t-16"
-              i18n="@@authorize.youAreCurrentlySignin"
-            >
-              You are currently signed in as:
-            </p>
-            <div class="m-t-small m-b-8">
-              <div class="mat-body-2">
-                <strong>{{ userName }}</strong>
-              </div>
-              <a
-                class="mat-button-font"
-                href="{{ orcidUrl }}"
-                target="_blank"
-                >{{ orcidUrl }}</a
-              >
-            </div>
-            <div c class="mat-body-2 user-links m-b-8">
-              <a
-                class="mat-button-font"
-                i18n="@@public-layout.sign_out"
-                (click)="signOut()"
-                >Sign out</a
-              >
-              <app-trusted-individuals-dropdown
-                *ngIf="trustedIndividuals"
-                (changeUser)="changeAccount($event)"
-                [trustedIndividuals]="trustedIndividuals"
-              ></app-trusted-individuals-dropdown>
-            </div>
-            <mat-divider class="orcid-divider m-b-16"></mat-divider>
-
-            <app-info-drop-down
-              class="m-t-16"
-              [name]="oauthRequest.clientName"
-              [description]="oauthRequest.clientDescription"
-            ></app-info-drop-down>
-
-            <p
-              class="mat-body-1 m-t-4 m-b-base"
-              i18n="@@authorize.hasAskedAccess"
-            >
-              has asked for the following access to your ORCID record:
-            </p>
-
-            <ul class="mat-body-1 m-t-base m-b-base">
-              <ng-container
-                *ngFor="let scope of this.oauthRequest.scopes; let last = last"
-              >
-                <li [matTooltip]="getLongDescription(scope)">
-                  <mat-divider></mat-divider>
-                  <div>
-                    <ng-container
-                      [ngTemplateOutlet]="
-                        (getIconName(scope) === 'updateIcon' &&
-                          updateIconTemplate) ||
-                        (getIconName(scope) === 'viewIcon' &&
-                          viewIconTemplate) ||
-                        (getIconName(scope) === 'orcidIcon' &&
-                          orcidIconTemplate)
-                      "
-                    ></ng-container>
-                    <div>{{ getDescription(scope) }}</div>
-                  </div>
-                  <mat-divider *ngIf="last"></mat-divider>
-                </li>
-              </ng-container>
-            </ul>
-
-            <button
-              mat-raised-button
-              color="primary"
-              i18n="@@authorize.authorize"
-              (click)="authorize()"
-            >
-              Authorize access
-            </button>
-            <button
-              mat-stroked-button
-              color="primary"
-              i18n="@@authorize.deny"
-              (click)="authorize(false)"
-            >
-              Deny access
-            </button>
-
-            <p class="mat-caption m-t-0">
-              <ng-container i18n="@@authorize.ifAuthorize">
-                If authorized, this organization will have access to your ORCID
-                record, as outlined above and described in further detail
-                in</ng-container
-              >
-
-              <a
-                href="https://orcid.org/footer/privacy-policy"
-                target="_blank"
-                rel="noopener noreferrer"
-                class="mat-button-font"
-                i18n="@@authorize.privacyPolicy"
-                >ORCID’s privacy policy.</a
-              >
-            </p>
-            <p class="mat-caption">
-              <ng-container i18n="@@authorize.youCanManage">
-                You can manage access permissions for this and other Trusted
-                Organizations in your
-              </ng-container>
-              <a
-                href="/account#manage-permissions"
-                target="_blank"
-                rel="noopener noreferrer"
-                class="mat-button-font"
-                i18n="@@authorize.accountSettings"
-                >account settings.</a
-              >
-            </p>
-          </mat-card-content>
-        </mat-card>
-      </div>
-    </div>
+    <app-form-authorize *ngIf="showAuthorizationComponent"></app-form-authorize>
+    <app-oauth-error *ngIf="!showAuthorizationComponent"></app-oauth-error>
   </div>
 </main>
-
-<ng-template #updateIconTemplate>
-  <mat-icon class="material-icons-outlined" role="presentation"
-    >update</mat-icon
-  >
-</ng-template>
-
-<ng-template #viewIconTemplate>
-  <mat-icon class="material-icons-outlined" role="presentation"
-    >remove_red_eye</mat-icon
-  >
-</ng-template>
-
-<ng-template #orcidIconTemplate>
-  <mat-icon role="presentation">
-    <img src="/assets/vectors/orcid.logo.black.icon.svg" alt=""
-  /></mat-icon>
-</ng-template>
