+++ src/app/core/oauth/oauth.service.ts
@@ -12,6 +12,7 @@ import {
   tap,
 } from 'rxjs/operators'
 import { WINDOW } from 'src/app/cdk/window'
+import { ApplicationRoutes } from 'src/app/constants'
 import { ERROR_REPORT } from 'src/app/errors'
 import { OauthParameters, RequestInfoForm } from 'src/app/types'
 import { OauthAuthorize } from 'src/app/types/authorize.endpoint'
@@ -27,6 +28,11 @@ const OAUTH_SESSION_ERROR_CODES_HANDLE_BY_CLIENT_APP = [
   'interaction_required',
   'invalid_scope',
   'unsupported_response_type',
+]
+
+const OAUTH_SESSION_ERROR_CODES_HANDLE_BY_ORCID_APP = [
+  'oauth_error',
+  'invalid_grant',
 ]
 
 @Injectable({
@@ -155,16 +161,9 @@ export class OauthService {
       this.window.location.href = `${session.redirectUrl}#error=${session.error}`
       return NEVER
     } else if (session.error || (session.errors && session.errors.length)) {
-      // Send the user to the signin and display a toaster error
-      let extra = {}
-      if (
-        session.error === 'oauth_error' ||
-        session.error === 'invalid_grant'
-      ) {
-        extra = { error: session.errorDescription }
-      }
+      // Send the user to the oauth page to see the error
       this._router
-        .navigate(['/signin'], { queryParams: extra })
+        .navigate([ApplicationRoutes.authorize])
         .then((navigated: boolean) => {
           if (navigated) {
             this._errorHandler
