+++ src/app/sign-in/pages/sign-in/sign-in.component.html
@@ -56,11 +56,6 @@
               [realUserOrcid]="realUserOrcid"
             >
             </app-logged-in>
-            <app-oauth-error
-              *ngIf="isOauthError"
-              [errorDescription]="errorDescription"
-            >
-            </app-oauth-error>
           </mat-card-content>
         </mat-card>
       </div>
